+++ src/hapi_schema/db_population.py
@@ -5,9 +5,10 @@ from datetime import datetime
 from sqlalchemy import (
     CheckConstraint,
     DateTime,
+    Enum,
     ForeignKey,
     Integer,
-    Text,
+    String,
     select,
     text,
 )
@@ -19,12 +20,15 @@ from hapi_schema.db_dataset import DBDataset
 from hapi_schema.db_location import DBLocation
 from hapi_schema.db_resource import DBResource
 from hapi_schema.utils.base import Base
+from hapi_schema.utils.enums import Gender
 from hapi_schema.utils.view_params import ViewParams
 
 
 class DBPopulation(Base):
     __tablename__ = "population"
     __table_args__ = (
+        CheckConstraint("min_age >= 0", name="min_age"),
+        CheckConstraint("max_age >= 0", name="max_age"),
         CheckConstraint("population >= 0", name="population"),
         CheckConstraint(
             "(reference_period_end >= reference_period_start) OR (reference_period_start IS NULL)",
@@ -32,20 +36,20 @@ class DBPopulation(Base):
         ),
     )
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    resource_hdx_id: Mapped[int] = mapped_column(
+    resource_hdx_id: Mapped[str] = mapped_column(
         ForeignKey("resource.hdx_id", onupdate="CASCADE", ondelete="CASCADE"),
         nullable=False,
     )
     admin2_ref: Mapped[int] = mapped_column(
-        ForeignKey("admin2.id", onupdate="CASCADE"), nullable=False
+        ForeignKey("admin2.id", onupdate="CASCADE"),
+        primary_key=True,
     )
-    gender_code: Mapped[str] = mapped_column(
-        ForeignKey("gender.code", onupdate="CASCADE"), nullable=True
+    gender: Mapped[Gender] = mapped_column(
+        Enum(Gender, name="gender_enum"), primary_key=True
     )
-    age_range_code: Mapped[str] = mapped_column(
-        ForeignKey("age_range.code", onupdate="CASCADE"), nullable=True
-    )
+    age_range: Mapped[str] = mapped_column(String(32), primary_key=True)
+    min_age: Mapped[int] = mapped_column(Integer, nullable=True, index=True)
+    max_age: Mapped[int] = mapped_column(Integer, nullable=True, index=True)
     population: Mapped[int] = mapped_column(
         Integer, nullable=False, index=True
     )
@@ -55,12 +59,9 @@ class DBPopulation(Base):
     reference_period_end: Mapped[datetime] = mapped_column(
         DateTime, nullable=True, server_default=text("NULL"), index=True
     )
-    source_data: Mapped[str] = mapped_column(Text, nullable=True)
 
     resource = relationship("DBResource")
     admin2 = relationship("DBAdmin2")
-    gender = relationship("DBGender")
-    age_range = relationship("DBAgeRange")
 
 
 view_params_population = ViewParams(