+++ tests/test_org.py
@@ -1,13 +1,13 @@
 from datetime import datetime
 
-from hdx.database.views import build_view
+from hdx.database import Database
 
 from hapi_schema.db_org import DBOrg, view_params_org
 
 
 def test_org_view(run_view_test):
     """Check that org view references org type."""
-    view_org = build_view(view_params_org.__dict__)
+    view_org = Database.prepare_view(view_params_org.__dict__)
     run_view_test(
         view=view_org,
         whereclause=(